refactoring activity: decoupling #279
