+++ client/src/app/services/transaction/transaction.service.ts
@@ -4,17 +4,18 @@ import { Transaction } from '../../interfaces/transaction';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { UserService } from '../user/user.service';
+import { dataFormatter } from '../../utilities/data-formatters'
 
 @Injectable({
   providedIn: 'root',
 })
 export class TransactionService {
-  constructor(private api: ApiService, private userService: UserService) {}
+  constructor(private api: ApiService, private userService: UserService, private dataFormatter: dataFormatter) {}
 
   getUserTransactions(userId: number): Observable<Transaction[]> {
     return this.api
       .get('transactions/' + userId)
-      .pipe(map((res: Response) => this.transactionListFormatter(res)));
+      .pipe(map((res: Response) => this.dataFormatter.transactionListFormatter(res)));
   }
 
   processStockTransaction(
@@ -23,23 +24,7 @@ export class TransactionService {
   ): Observable<any> {
     return this.api
       .post('transactions/execute/' + userId, transaction)
-      .pipe(map((res: Response) => this.transactionListFormatter(res)));
+      .pipe(map((res: Response) => this.dataFormatter.transactionListFormatter(res)));
   }
 
-  transactionListFormatter(response: any): Transaction[] {
-    const result: Transaction[] = [];
-    for (const fetchedTransaction of response.body) {
-      result.push({
-        qty: fetchedTransaction.quantity,
-        time_in_force: fetchedTransaction.purchasedAt,
-        type: fetchedTransaction.moneyTreeOrderType,
-        client_order_id: fetchedTransaction.clienOrderId,
-        status: fetchedTransaction.status,
-        averagePricePerShare: fetchedTransaction.avgPrice,
-        symbol: fetchedTransaction.symbol,
-        total: fetchedTransaction.total,
-      });
-    }
-    return result;
-  }
 }
